py/requirements.txt
mysql-connector-python
markdown
beautifulsoup4